[CORE-656] Update Dashboard Team layout (#5798)
## Problem solved

CORE-656

* Move the Team analytics from overview page to Analytics tab
* Move the changelog to "projects" page and make that the default page to show on /team/<team-slug>

<!-- start pr-codex -->

---

## PR-Codex overview
This PR primarily focuses on updating links related to the `thirdweb dashboard`, changing references from project-specific URLs to a more general team URL. Additionally, it refines the `TeamProjectsPage` component and enhances the analytics page structure.

### Detailed summary
- Updated links in multiple `.mdx` files to point to `https://thirdweb.com/team` instead of project-specific URLs.
- Modified `layout.tsx` to change the path from `projects` to `analytics`.
- Enhanced `TeamProjectsPage` to display a message when no projects are created, with a button to create a new project.
- Refactored the main page to include the `TeamProjectsPage` component and removed unnecessary analytics content.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
