+++ apps/portal/src/app/account/api-keys/access/page.mdx
@@ -26,7 +26,7 @@ To modify the allowed Domain Id or Bundle Ids:
 
 <Step title="Go to Dashboard">
 
-Go to **[thirdweb dashboard](https://thirdweb.com/team/~/~/projects)**
+Go to **[thirdweb dashboard](https://thirdweb.com/team)**
 
 </Step>
