+++ apps/portal/src/app/account/api-keys/create/page.mdx
@@ -17,7 +17,7 @@ export const metadata = createMetadata({
 
 <Step title="Login to dashboard">
 
-Log in to the dashboard and go to your **[Team projects page](https://thirdweb.com/team/~/~/projects)**
+Log in to the dashboard and go to your **[Team projects page](https://thirdweb.com/team)**
 
 </Step>
