+++ apps/portal/src/app/account/api-keys/delete/page.mdx
@@ -20,7 +20,7 @@ Deleting an API key will invalidate it, making it no longer usable.
 <Steps>
 
 <Step title="Go to Dashboard">
-	Go to **[your team's projects page](https://thirdweb.com/team/~/~/projects/)** on the dashboard.
+	Go to **[your team's projects page](https://thirdweb.com/team)** on the dashboard.
 </Step>
 
 <Step title='Choose API Key'>
