+++ apps/portal/src/app/account/api-keys/edit-services/page.mdx
@@ -23,7 +23,7 @@ All services on API keys are enabled by default. If you want to disable any serv
 
 <Step title="Go to Dashboard">
 
-Go to **[your team's projects page](https://thirdweb.com/team/~/~/projects)** on the dashboard.
+Go to **[your team's projects page](https://thirdweb.com/team)** on the dashboard.
 
 </Step>
