+++ apps/portal/src/app/connect/pay/fee-sharing/page.mdx
@@ -18,6 +18,6 @@ For example, if a user purchases $100 worth of Polygon through your application,
 
 Plase note that fee sharing only applies to fees collected through swaps and bridges. Fiat purchases do not currently qualify for fee sharing.
 
-To participate in fee sharing, all you need to do is set a recipient address in your dashboard. Your recipient address can be configured in your dashboard by navigating to your Team > Project > Connect > Pay in [thirdweb dashboard](https://www.thirdweb.com/team/~/~/projects).
+To participate in fee sharing, all you need to do is set a recipient address in your dashboard. Your recipient address can be configured in your dashboard by navigating to your Team > Project > Connect > Pay in [thirdweb dashboard](https://www.thirdweb.com/team).
 
 <DocImage src={paySettingsImage} />
