+++ apps/portal/src/app/connect/pay/guides/build-a-custom-experience/page.mdx
@@ -41,7 +41,7 @@ In this guide, we'll show you how to purchase 0.01 Base ETH from USD in Typescri
 />
 <Step title='Get Your Client ID'>
 
-Log in to the [thirdweb dashboard](https://thirdweb.com/team/~/~/projects). Click on Create New > Project to get your **Client ID**. You'll need your Client ID to interact with the Connect SDK.
+Log in to the [thirdweb dashboard](https://thirdweb.com/team). Click on Create New > Project to get your **Client ID**. You'll need your Client ID to interact with the Connect SDK.
 
 </Step>
 </Step>
