+++ apps/portal/src/app/connect/quickstart/page.mdx
@@ -17,7 +17,7 @@ npm i thirdweb
 
 <Step title="Get Your Client ID">
 
-Log in to the [thirdweb dashboard](https://thirdweb.com/team/~/~/projects). Create a new project to get your **Client ID**.
+Log in to the [thirdweb dashboard](https://thirdweb.com/team). Create a new project to get your **Client ID**.
 
 </Step>
