+++ apps/portal/src/app/contracts/modular-contracts/get-started/deploy-modular-contract/page.mdx
@@ -34,7 +34,7 @@ as this guide will build on top of those previous guides.
 <Steps>
 
 <Step title="Grab an API Key from the thirdweb Dashboard">
-Go to the [thirdweb Dashboard](https://thirdweb.com/team/~/~/projects), Create a new Project to obtain API Key
+Go to the [thirdweb Dashboard](https://thirdweb.com/team), Create a new Project to obtain API Key
 
 <DocImage src={ createApiKey } />
