+++ apps/portal/src/app/infrastructure/rpc-edge/overview/page.mdx
@@ -21,7 +21,7 @@ Remote Procedure Call (RPC) Edge provides reliable access to querying data and i
 By default, we provide publicly available RPCs for over 900+ EVM Networks. [View the default RPC endpoints](https://thirdweb.com/chainlist) for each supported chain.
 
 <Callout variant="info" title="Accessing your RPC URL">
-	To use RPC Edge with other tools like Anvil, take a chain's [public RPC endpoint](https://thirdweb.com/chainlist) and attach your app's client ID from Project > Settings page in [dashboard](https://thirdweb.com/team/~/~/projects) `https://<chainId>.rpc.thirdweb.com/<clientId>`.
+	To use RPC Edge with other tools like Anvil, take a chain's [public RPC endpoint](https://thirdweb.com/chainlist) and attach your app's client ID from Project > Settings page in [dashboard](https://thirdweb.com/team) `https://<chainId>.rpc.thirdweb.com/<clientId>`.
 </Callout>
 
 #### Features
