+++ apps/portal/src/app/payments/nft-checkout/api-reference/page.mdx
@@ -12,7 +12,7 @@ The Payments API enables all the functionalities needed to create checkouts.
 [View the full API reference.](https://redocly.github.io/redoc/?url=https://payments.thirdweb.com/api/doc)
 
 You must use an API Secret Key to make authenticated calls. These calls should be
-made from your backend. [Create an API Key by creating a project from your thirdweb dashboard.](https://thirdweb.com/team/~/~/projects)
+made from your backend. [Create an API Key by creating a project from your thirdweb dashboard.](https://thirdweb.com/team)
 
 Provide your secret key as a header: `x-secret-key: your_api_secret_key`
