+++ apps/portal/src/app/react-native/v0/getting-started/page.mdx
@@ -18,7 +18,7 @@ Our SDK uses a [Provider Pattern](https://flexiple.com/react/provider-pattern-wi
 
 You will require an API key to use thirdweb's infrastructure services with the SDK.
 
-you need to first obtain an API key from the [dashboard](https://thirdweb.com/team/~/~/projects) by creating a new project and then copy the "Client ID" and pass it to the `ThirdwebPovider` as the `clientId` prop.
+you need to first obtain an API key from the [dashboard](https://thirdweb.com/team) by creating a new project and then copy the "Client ID" and pass it to the `ThirdwebPovider` as the `clientId` prop.
 
 </Callout>
