+++ apps/portal/src/app/react/v4/ThirdwebProvider/page.mdx
@@ -26,7 +26,7 @@ The `ThirdwebProvider` is a wrapper component that provides access to all of the
 
 You will require an API key to use thirdweb's infrastructure services with the SDK.
 
-you need to first obtain an API key from the [dashboard](https://thirdweb.com/team/~/~/projects) by creating a new project and then copy the "Client ID" and pass it to the `ThirdwebProvider` as the `clientId` prop.
+you need to first obtain an API key from the [dashboard](https://thirdweb.com/team) by creating a new project and then copy the "Client ID" and pass it to the `ThirdwebProvider` as the `clientId` prop.
 
 </Callout>
