+++ apps/portal/src/app/react/v5/rainbow-kit-migrate/page.mdx
@@ -85,7 +85,7 @@ Learn how to migrate to thirdweb's ConnectButton component from [RainbowKit](htt
       );
     };
     ```
-    <Callout variant="info" title="Client ID">Get a free client ID to use in your application by creating a project in [thirdweb's dashboard](https://thirdweb.com/team/~/~/projects).</Callout>
+    <Callout variant="info" title="Client ID">Get a free client ID to use in your application by creating a project in [thirdweb's dashboard](https://thirdweb.com/team).</Callout>
 	</Step>
   <Step title="Customize">
     To customize your ConnectButton component, view the cheatsheet below or [view the customization documentation](https://portal.thirdweb.com/connect/sign-in/customization).
